refactor: create reusable search component for unsplash
